+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ## Bug Fixes
  * JDBC regression- `PreparedStatement#executeUpdate()` and `PreparedStatement#executeLargeUpdate()` are not traced (#918)
  * When systemd cgroup driver is used, the discovered Kubernetes pod UID contains "_" instead of "-" (#920)
+ * DB2 jcc4 driver is not traced properly (#926)
 
 # 1.11.0